refactor(cli): establish command module ownership
