+++ app/src/cli/commands/admin/handler.rs
@@ -1,4 +1,4 @@
-use super::{output, prompt};
+use crate::cli::{output, prompt};
 use crate::{
   cli::args::AdminCommand,
   config::{ServerConfig, ServerOverrides, database_path, ensure_data_dir},
@@ -12,7 +12,7 @@ use std::{
 };
 use zip::{CompressionMethod, ZipWriter, write::SimpleFileOptions};
 
-pub(super) async fn execute(
+pub(crate) async fn execute(
   command: AdminCommand,
   data_dir: Option<PathBuf>,
   json_output: bool,
