+++ app/src/cli/commands/auth/mod.rs
@@ -1,0 +1,3 @@
+mod handler;
+
+pub(super) use handler::{login, logout};
