+++ app/src/cli/commands/environment/mod.rs
@@ -1,0 +1,3 @@
+mod handler;
+
+pub(super) use handler::{env_id, execute, resolve_environment};
