+++ app/src/cli/commands/secret/mod.rs
@@ -1,0 +1,3 @@
+mod handler;
+
+pub(super) use handler::execute;
