+++ app/src/cli/commands/server/mod.rs
@@ -1,0 +1,3 @@
+mod handler;
+
+pub(super) use handler::execute;
