+++ app/src/cli/commands/update/handler.rs
