+++ app/src/cli/output.rs
